Refactor XML component imports: streamline import statements across multiple files for improved clarity and organization
