+++ spinta/config.py
@@ -77,7 +77,7 @@ CONFIG = {
             "sqldump": "spinta.datasets.backends.sqldump.components:SqlDump",
             "dask": "spinta.datasets.backends.dataframe.components:DaskBackend",
             "dask/csv": "spinta.datasets.backends.dataframe.backends.csv.components:Csv",
-            "dask/xml": "spinta.datasets.backends.dataframe.backends.xml.components.xml:Xml",
+            "dask/xml": "spinta.datasets.backends.dataframe.backends.xml.components:Xml",
             "dask/json": "spinta.datasets.backends.dataframe.backends.json.components:Json",
             "dask/memory": "spinta.datasets.backends.dataframe.backends.memory.components:MemoryDaskBackend",
             "xlsx": "spinta.datasets.backends.notimpl.components:BackendNotImplemented",
