+++ spinta/datasets/backends/dataframe/backends/xml/commands/read.py
@@ -9,11 +9,12 @@ from lxml import etree
 from spinta import commands
 from spinta.components import Context, Model, Property
 from spinta.core.ufuncs import Expr
-from spinta.datasets.backends.dataframe.backends.xml.components.row_meta_item import RowMetaItem
-from spinta.datasets.backends.dataframe.backends.xml.components.row import Row
-from spinta.datasets.backends.dataframe.backends.xml.components.dask_xml import DaskXml
-from spinta.datasets.backends.dataframe.backends.xml.components.row_formatter import RowFormatter
-from spinta.datasets.backends.dataframe.backends.xml.components import Xml
+from spinta.datasets.backends.dataframe.backends.xml.components import (
+    RowMetaItem,
+    Row,
+    DaskXml,
+    Xml,
+)
 from spinta.datasets.backends.dataframe.commands.read import (
     parametrize_bases,
     get_dask_dataframe_meta,
